Refactor image compression logic into `ImageCompressorHelper`, update resume processing to handle file uploads, and clean up redundant methods in services.
